// apps/bridge/codegen/model/acrossAvailableRouteDto.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * API
 * API docs
 * OpenAPI spec version: 1.0
 */

export interface AcrossAvailableRouteDto {
  destinationChainId: number;
  destinationToken: string;
  destinationTokenSymbol: string;
  originChainId: number;
  originToken: string;
  originTokenSymbol: string;
}